TaskPageもmount追加してmountコードを共通関数化
