+++ components/ModalTasks.vue
@@ -54,7 +54,7 @@
 </template>
 
 <script setup lang="ts">
-import { ref, defineProps, defineEmits } from 'vue';
+import { ref, defineProps } from 'vue';
 import { Task, TaskManager } from '~/types/types';
 
 const props = defineProps({
