+++ components/TodoList.vue
@@ -83,12 +83,15 @@ const dropCard = (event: any, status: 'todo' | 'doing' | 'done') => {
   const index = taskManagers.value.findIndex(
     (taskManager) => taskManager.companyId === Number(dragCompanyId)
   );
+  console.log(index);
   if (index !== -1) {
     taskManagers.value[index].tasks.findIndex((task) => {
       if (task.taskId === Number(dragTaskId)) {
         task.state = status;
+        console.log('task.state', task.state);
       }
     });
   }
+  console.log(taskManagers.value);
 };
 </script>
