+++ pages/task.vue
@@ -31,4 +31,6 @@ const list3 = <TodoListType>{
   cards: done,
   status: 'done',
 };
+
+monitaringUserData();
 </script>
